Add cherry picked VerifyingKey serialization
Here's the original PR by nalinbhardwaj:
https://github.com/zcash/halo2/pull/661

It was decided not to merge this PR, but at the moment this is the only way to serialize Halo2 VerifyingKey, since PSE's fork of Halo2 only supports serializing keys derived for KZG based proofs which were incompatible with the Poseidon gadget.

This branch should not be merged and is simple used as a temporary solution until halo2 implements proper serialization.